shader renaming
