+++ src/point-preprocessor/tileDepthKey.wgsl
