+++ src/splatting-web/shaders/rasterize.wgsl
