+++ src/splatting-web/shaders/rasterizeSharedMemory.wgsl
