+++ src/splatting-web/shaders/screen.wgsl
